chore: move `Sends` files to interface folder
